+++ ft-sdk/src/data.rs
@@ -4,6 +4,7 @@ pub type Result = std::result::Result<ft_sdk::chr::CHR<Output>, ft_sdk::Error>;
 pub enum Output {
     Json(serde_json::Value),
     Binary(Binary),
+    Redirect(String),
 }
 
 #[derive(Debug)]
@@ -16,7 +17,7 @@ pub struct Binary {
 pub(crate) fn binary_response(
     binary: Binary,
 ) -> std::result::Result<::http::Response<bytes::Bytes>, ft_sdk::Error> {
-    let mut response_builder = ::http::Response::builder()
+    let mut response_builder = http::Response::builder()
         .status(200)
         .header("Content-Type", binary.content_type.as_str());
 
@@ -44,12 +45,15 @@ impl From<ft_sdk::chr::CHR<Output>>
         let response = match response {
             Output::Json(json_value) => ft_sdk::json(json_value),
             Output::Binary(binary) => binary_response(binary),
+            Output::Redirect(url) => Ok(http::Response::builder()
+                .status(200)
+                .body(format!("<meta http-equiv='refresh' content='0; url={url}' />").into())?),
         }?;
         ft_sdk::chr::chr(cookies, headers, response)
     }
 }
 
-/// Creates a response that instructs browser to store in downloads the binary content provided.
+/// Creates a response that instructs the browser to store in downloads the binary content provided.
 ///
 /// # Parameters
 /// - `filename`: An optional `String` representing the name of the file. If provided, the response
@@ -98,6 +102,26 @@ pub fn binary<S: AsRef<str>>(content: bytes::Bytes, content_type: S) -> Result {
         content,
         content_type: content_type.as_ref().to_string(),
     })))
+}
+
+/// Returns a redirect using 200 response. This is used for setting cookie while redirect.
+///
+/// Adding cookie with redirect headers does not work across browsers. This helper creates
+/// a 200 OK response, with a meta-refresh tag to redirect the browser.
+///
+/// ```rust
+///  let cookie = cookie::Cookie::build((ft_sdk::auth::SESSION_KEY, sid))
+///         .domain(host.without_port())
+///         .path("/")
+///         .max_age(cookie::time::Duration::seconds(34560000))
+///         .same_site(cookie::SameSite::Strict)
+///         .build();
+/// Ok(ft_sdk::data::redirect("/").unwrap().with_cookie(cookie))
+/// ```
+pub fn redirect<S: AsRef<str>>(url: S) -> Result {
+    Ok(ft_sdk::chr::CHR::new(Output::Redirect(
+        url.as_ref().to_string(),
+    )))
 }
 
 pub fn json<T: serde::Serialize>(t: T) -> Result {